+++ src/components/orderList/OrderListItem.jsx
@@ -3,6 +3,7 @@ import styled from 'styled-components';
 import PropTypes from 'prop-types';
 import { COLOR } from '../../constants/color';
 import { Button, ProductImage, BUTTON_TYPE, PRODUCT_IMAGE_TYPE } from '..';
+import useInsertingItemToShoppingCart from '../../hooks/useInsertingItemToShoppingCart';
 
 const Container = styled.div`
   display: flex;
@@ -22,22 +23,33 @@ const Info = styled.div`
   color: ${COLOR.GRAY_600};
 `;
 
-const OrderListItem = ({ src, alt, name, quantity, price }) => (
-  <Container>
-    <ProductImage type={PRODUCT_IMAGE_TYPE.SMALL} src={src} alt={alt} />
-    <TextWrapper>
-      <Name>{name}</Name>
-      <Info>
-        {price.toLocaleString('ko-KR')}원 / 수량: {quantity}개
-      </Info>
-    </TextWrapper>
-    <div style={{ marginLeft: 'auto' }}>
-      <Button type={BUTTON_TYPE.SMALL}>장바구니</Button>
-    </div>
-  </Container>
-);
+const OrderListItem = ({ id, src, alt, name, quantity, price }) => {
+  const { insertShoppingCart, isDialogOpen, Dialog } = useInsertingItemToShoppingCart({ product_id: id });
+
+  return (
+    <>
+      <Container>
+        <ProductImage type={PRODUCT_IMAGE_TYPE.SMALL} src={src} alt={alt} />
+        <TextWrapper>
+          <Name>{name}</Name>
+          <Info>
+            {price.toLocaleString('ko-KR')}원 / 수량: {quantity}개
+          </Info>
+        </TextWrapper>
+        <div style={{ marginLeft: 'auto' }}>
+          <Button onClick={insertShoppingCart} type={BUTTON_TYPE.SMALL}>
+            장바구니
+          </Button>
+        </div>
+      </Container>
+
+      {isDialogOpen && <Dialog />}
+    </>
+  );
+};
 
 OrderListItem.propTypes = {
+  id: PropTypes.number.isRequired,
   src: PropTypes.string.isRequired,
   alt: PropTypes.string.isRequired,
   name: PropTypes.string.isRequired,
