+++ src/components/productList/ProductListItem.jsx
@@ -3,10 +3,9 @@ import styled from 'styled-components';
 import PropTypes from 'prop-types';
 import shoppingCartImg from '../../assets/shoppingCart.svg';
 import { useHistory } from 'react-router';
-import useDialog from '../../hooks/useDialog';
-import { ProductImage, Dialog, PRODUCT_IMAGE_TYPE, DIALOG_TYPE } from '..';
+import { ProductImage, PRODUCT_IMAGE_TYPE } from '..';
 import { PATH } from '../../constants/path';
-import useInsertingShoppingCart from '../../hooks/useInsertingShoppingCart';
+import useInsertingItemToShoppingCart from '../../hooks/useInsertingItemToShoppingCart';
 
 const Content = styled.ul`
   display: flex;
@@ -36,9 +35,10 @@ const Image = styled.img`
 `;
 
 const ProductListItem = ({ id, name, src, price }) => {
-  const { isDialogOpen, setIsDialogOpen, clickConfirm, clickCancel, type, setType } = useDialog();
-  const { insertShoppingCart, ADD_FAILURE } = useInsertingShoppingCart({ product_id: id, setIsDialogOpen, setType });
   const history = useHistory();
+  const { insertShoppingCart, isDialogOpen, Dialog } = useInsertingItemToShoppingCart({
+    product_id: id,
+  });
 
   const handleRoutingProductDetail = () => {
     history.push(PATH.PRODUCT_DETAIL, { id });
@@ -61,11 +61,7 @@ const ProductListItem = ({ id, name, src, price }) => {
         </Content>
       </div>
 
-      {isDialogOpen && (
-        <Dialog type={DIALOG_TYPE.ALERT} onConfirm={clickConfirm} onClose={clickCancel}>
-          {type === ADD_FAILURE ? '이미 장바구니에 추가되어 있습니다.' : '장바구니에 추가되었습니다.'}
-        </Dialog>
-      )}
+      {isDialogOpen && <Dialog />}
     </>
   );
 };
